requirements.txt
pyyaml
pytest
requests
jsonpath